requirements.txt
pydf2
selenium
beautifulsoup4
requests
google-generativeai